+++ app/src/main/java/de/uni/hannover/hci/mi/team6/covidcheckin/MainActivity.kt
@@ -2,7 +2,6 @@ package de.uni.hannover.hci.mi.team6.covidcheckin
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import de.uni.hannover.hci.mi.team6.covidcheckin.ui.main.ExportSheetFragment
 
 class MainActivity : AppCompatActivity() {
 
@@ -11,7 +10,8 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.main_activity)
         if (savedInstanceState == null) {
 
-            val exportSheetFragment = ExportSheetFragment();
+            val exportSheetFragment =
+                ExportSheetFragment();
             exportSheetFragment.show(supportFragmentManager, "ExportSheet")
 
         }
